Remove die() from controller rendering
I always thought it was a little strange that the controller had to call
die().  Now I actually have a need to avoid it: automated testing.  If
the controller assumes that there is nothing left to do, we can't do
integrated testing.  This also adds a check to make sure that the
filter 'template_include' only returns something once.  This may be a
strange solution, but it allows for testing because die() doesn't have
to be called anymore.