Restore die() but add conditional switch
Turns out die()ing is necessary a lot of the time.  I added a property to
the MvcController class so it is easy to turn off die()ing when you need
to (for testing), but the default behavior is to die().  Turning it off
can happen through the dispatcher or by setting ->should_die to
false.